move Pollution class to pollution package
