+++ src/main/java/gregtech/common/misc/GTCommand.java
@@ -25,7 +25,7 @@ import gregtech.api.enums.GTValues;
 import gregtech.api.objects.GTChunkManager;
 import gregtech.api.util.GTMusicSystem;
 import gregtech.api.util.GTUtility;
-import gregtech.common.Pollution;
+import gregtech.common.pollution.Pollution;
 import gregtech.common.misc.spaceprojects.SpaceProjectManager;
 
 public final class GTCommand extends CommandBase {
