+++ src/main/java/gregtech/common/tileentities/machines/multi/MTECharcoalPit.java
@@ -35,7 +35,7 @@ import gregtech.api.recipe.check.CheckRecipeResultRegistry;
 import gregtech.api.render.TextureFactory;
 import gregtech.api.util.MultiblockTooltipBuilder;
 import gregtech.api.util.WorldSpawnedEventBuilder;
-import gregtech.common.Pollution;
+import gregtech.common.pollution.Pollution;
 
 public class MTECharcoalPit extends MTETooltipMultiBlockBase implements ISecondaryDescribable {
