+++ src/mixin/java/gregtech/mixin/mixins/late/ic2/MixinIC2IronFurnacePollution.java
@@ -8,7 +8,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import gregtech.common.Pollution;
+import gregtech.common.pollution.Pollution;
 import gregtech.common.pollution.PollutionConfig;
 import ic2.core.block.machine.tileentity.TileEntityIronFurnace;
