+++ src/mixin/java/gregtech/mixin/mixins/late/railcraft/MixinRailcraftTunnelBorePollution.java
@@ -9,7 +9,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import gregtech.common.Pollution;
+import gregtech.common.pollution.Pollution;
 import gregtech.common.pollution.PollutionConfig;
 import mods.railcraft.common.carts.EntityTunnelBore;
